Introduce ChangeFeedListener to ensure correct order for HostedServices
