+++ src/Atc.Cosmos/Internal/IChangeFeedListener.cs
@@ -1,0 +1,12 @@
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace Atc.Cosmos.Internal
+{
+    public interface IChangeFeedListener
+    {
+        Task StartAsync(CancellationToken cancellationToken);
+
+        Task StopAsync(CancellationToken cancellationToken);
+    }
+}
